bookcontroller fait (normalement)

+++ src/controllers/book.controller.ts
@@ -24,24 +24,37 @@ import { checkPermissions } from "../middlewares/permissions";
 export class BookController extends Controller {
   @Get("/")
   public async getAllBooks(): Promise<BookOutputDTO[]> {
+    if(checkPermissions("book", "read")){
     return bookService.getAllBooks();
+    }
+    else{
+      throw new Error("Forbidden: Insufficient permissions to create a book");
+    }
   }
 
   @Get("{id}")
   public async getBook(@Path("id") id: number): Promise<BookOutputDTO> {
+    if(checkPermissions("book", "read")){
     return await bookService.getBookById(id);
+  }else{
+    throw new Error("Forbidden: Insufficient permissions to create a book");
+  }
   }
 
   @Post("/")
   public async postBooks(
     @Body() requestBody: BookInputDTO,
   ): Promise<BookOutputDTO> {
-    return bookService.createBook(
-      requestBody.title,
-      requestBody.publish_year,
-      requestBody.author_id,
-      requestBody.isbn,
-    );
+    if(checkPermissions("book", "write")){
+      return bookService.createBook(
+        requestBody.title,
+        requestBody.publish_year,
+        requestBody.author_id,
+        requestBody.isbn,
+      );
+    }else{
+      throw new Error("Forbidden: Insufficient permissions to create a book");
+    }
   }
 
   
@@ -50,6 +63,7 @@ export class BookController extends Controller {
     @Path("id") id: number,
     @Body() requestBody: BookInputPatchDTO,
   ): Promise<BookOutputDTO> {
+    if(checkPermissions("book", "write")){
     return bookService.updateBook(
       id,
       requestBody.title,
@@ -57,17 +71,28 @@ export class BookController extends Controller {
       requestBody.author_id,
       requestBody.isbn,
     );
+    }else{
+      throw new Error("Forbidden: Insufficient permissions to create a book");
+    }
   }
 
   @Delete("{id}")
   public async deleteBook(@Path("id") id: number): Promise<void> {
+    if(checkPermissions("book", "delete")){
     await bookService.deleteBook(id);
+    }else{
+      throw new Error("Forbidden: Insufficient permissions to create a book");
+    }
   }
 
   @Get("{id}/book-collections")
   public async getBookCollectionsByBookId(
     @Path() id: number,
   ): Promise<BookCollectionOutputDTO[]> {
+    if(checkPermissions("book", "read")){
     return bookService.getBookCollectionsByBookId(id);
+  }else{
+    throw new Error("Forbidden: Insufficient permissions to create a book");
   }
+}
 }
